+++ packages/featureserver/lib/constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+  CURRENT_VERSION: 10.51,
+  FULL_VERSION: '10.5.1'
+}