+++ packages/featureserver/lib/defaults/index.js
@@ -1,4 +1,3 @@
 module.exports = {
-  serverMetadata: require('./server-metadata'),
-  version: require('./version')
+  serverMetadata: require('./server-metadata')
 }