+++ packages/featureserver/lib/defaults/server-metadata.js
@@ -1,6 +1,4 @@
-const version = require('./version')
 const serviceResponseDefaults = {
-  ...version,
   serviceDescription: 'This is a feature service exposed with Koop, an open source project that turns APIs into features. Service Description information may not be available for all services. For more information, check out https://github.com/koopjs/koop.',
   hasVersionedData: false,
   supportsDisconnectedEditing: false,