+++ packages/featureserver/lib/rest-info-route-handler.js
@@ -1,8 +1,18 @@
-const restInfoTemplate = require('../templates/rest-info.json')
-const versionTemplate = require('../templates/version.json')
+const _ = require('lodash')
+const {
+  CURRENT_VERSION,
+  FULL_VERSION
+} = require('./constants')
 
-function restInfo (dataSourceRestInfo) {
-  return { ...restInfoTemplate, ...versionTemplate, ...dataSourceRestInfo }
+function restInfo (data = {}, req) {
+  const currentVersion = _.get(req, 'app.locals.config.featureServer.currentVersion', CURRENT_VERSION)
+  const fullVersion = _.get(req, 'app.locals.config.featureServer.fullVersion', FULL_VERSION)
+
+  return {
+    currentVersion,
+    fullVersion,
+    ...data
+  }
 }
 
 module.exports = restInfo