+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- override default currentVersion and fullVersion via app.locals.config.featureServer
+
 ## [3.5.1] - 06-01-2022
 ### Changed
 - bump winnow version